requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
pymupdf==1.24.3
python-docx==1.1.2
pytesseract==0.3.10
Pillow==10.3.0
spacy>=3.8.0
# After installing, run: python -m spacy download en_core_web_sm
httpx>=0.27.0
aiofiles>=23.0.0
